+++ helpers/buttons_sbd_menu_theme.js
@@ -1,5 +1,6 @@
 ﻿include(fb.ProfilePath + 'scripts\\SMP\\xxx-scripts\\helpers\\menu_xxx.js');
 include(fb.ProfilePath + 'scripts\\SMP\\xxx-scripts\\helpers\\helpers_xxx_file.js');
+include(fb.ProfilePath + 'scripts\\SMP\\xxx-scripts\\helpers\\helpers_xxx_tags.js');
 
 const themeMenu = new _menu();
 
@@ -20,10 +21,49 @@ function createThemeMenu(parent) {
 	// Header
 	themeMenu.newEntry({entryText: 'Set theme file:', func: null, flags: MF_GRAYED});
 	themeMenu.newEntry({entryText: 'sep'});
+	themeMenu.newEntry({entryText: 'Open themes folder', func: () => {
+		if (_isFile(properties.theme[1])) {_explorer(properties.theme[1]);} // Open current file
+		else {_explorer(fb.ProfilePath + 'scripts\\SMP\\xxx-scripts\\presets\\Search by\\themes');} // or folder
+	}});
+	// Create theme
+	themeMenu.newEntry({entryText: 'Create theme file with selected track', func: () => {
+		// Tag names
+		const genreTag = properties.genreTag[1].split(',').filter(Boolean);
+		const styleTag = properties.styleTag[1].split(',').filter(Boolean);
+		const moodTag = properties.moodTag[1].split(',').filter(Boolean);
+		const dateTag = properties.dateTag[1].split(',').filter(Boolean); // only allows 1 value, but put it into an array
+		const composerTag = properties.composerTag[1].split(',').filter(Boolean);
+		const customStrTag = properties.customStrTag[1].split(',').filter(Boolean);
+		const customNumTag = properties.customNumTag[1].split(',').filter(Boolean); // only allows 1 value, but put it into an array
+		// Tag Values
+		const selHandleList = new FbMetadbHandleList(fb.GetFocusItem());
+		const genre = genreTag.length ? getTagsValuesV3(selHandleList, genreTag, true).flat().filter(Boolean) : [];
+		const style = styleTag.length ? getTagsValuesV3(selHandleList, styleTag, true).flat().filter(Boolean) : [];
+		const mood = moodTag.length ? getTagsValuesV3(selHandleList, moodTag, true).flat().filter(Boolean) : [];
+		const composer = composerTag.length ? getTagsValuesV3(selHandleList, composerTag, true).flat().filter(Boolean) : [];
+		const customStr = customStrTag.length ? getTagsValuesV3(selHandleList, customStrTag, true).flat().filter(Boolean) : [];
+		const restTagNames = ['key', dateTag.length ? dateTag[0] : 'skip', 'bpm', customNumTag.length ? customNumTag[0] : 'skip']; // 'skip' returns empty arrays...
+		const [keyArr, dateArr, bpmArr, customNumArr] = getTagsValuesV4(selHandleList, restTagNames).flat();
+		const key = keyArr;
+		const date = dateTag.length ? [Number(dateArr[0])] : [];
+		const bpm = bpmArr.length ? [Number(bpmArr[0])] : [];
+		const customNum = customNumTag.length ? [Number(customNumArr[0])] : [];
+		// Theme obj
+		let input = '';
+		try {input = utils.InputBox(window.ID, 'Enter theme name', 'Search by distance', 'my theme', true);}
+		catch (e) {return;}
+		if (!input.length) {return;}
+		const theme = {name: input, tags: []};
+		theme.tags.push({genre, style, mood, key, date, bpm, composer, customStr, customNum});
+		const filePath = fb.ProfilePath + 'scripts\\SMP\\xxx-scripts\\presets\\Search by\\themes\\' + input + '.json';
+		const bDone = _save(filePath, JSON.stringify(theme, null, '\t'));
+		if (!bDone) {fb.ShowPopupMessage('Error saving theme file:' + filePath, 'Search by distance'); return;}
+	}, flags: fb.GetFocusItem(true) ? MF_STRING : MF_GRAYED});
+	themeMenu.newEntry({entryText: 'sep'});
 	themeMenu.newEntry({entryText: 'None', func: () => {
 		properties.theme[1] = '';
 		const themeName = forcedTheme ? forcedTheme.name + ' (forced by recipe)' : 'None'; // Recipe may overwrite theme
-		parent.description = 'Search according to variables at properties.\n(Shift + L. Click to set theme)\t -> ' + themeName + '\n(Ctrl + L. Click to set recipe)\t -> ' + data.recipe;
+		parent.description = 'Search according to variables at properties.\n(Shift + L. Click to set theme)\t -> ' + themeName + '\n(Ctrl + L. Click to set recipe)\t -> ' + data.recipe + '\n(Shift + Ctrl + L. Click to set other config)';
 		data.tooltip = parent.description;
 		data.theme = 'None'; // Recipe may overwrite theme
 		properties.data[1] = JSON.stringify(data);
@@ -44,8 +84,8 @@ function createThemeMenu(parent) {
 			console.log('File is not a valid theme: ' + (theme.hasOwnProperty('tags') && tagCheck !== -1 ? [...new Set(tagsToCheck).difference(new Set(Object.keys(theme.tags[tagCheck])))] : file));
 			return;
 		}
-		// List
-		options.push(file);
+		// List files, with full path or relative path (portable)
+		options.push(_isFile(fb.FoobarPath + 'portable_mode_enabled') && file.indexOf(fb.ProfilePath) !== -1 ? file.replace(fb.ProfilePath,'.\\profile\\') : file);
 	});
 	const menus = [];
 	options.forEach((file) => {
@@ -56,7 +96,7 @@ function createThemeMenu(parent) {
 		menus.push(entryText);
 		themeMenu.newEntry({entryText, func: () => {
 			properties.theme[1] = file;
-			parent.description = 'Search according to variables at properties.\n(Shift + L. Click to set theme)\t -> ' + name + '\n(Ctrl + L. Click to set recipe)\t -> ' + data.recipe;
+			parent.description = 'Search according to variables at properties.\n(Shift + L. Click to set theme)\t -> ' + name + '\n(Ctrl + L. Click to set recipe)\t -> ' + data.recipe + '\n(Shift + Ctrl + L. Click to set other config)';
 			data.tooltip = parent.description;
 			data.theme = theme.name;
 			properties.data[1] = JSON.stringify(data);
@@ -67,9 +107,5 @@ function createThemeMenu(parent) {
 		const idx = options.indexOf(properties.theme[1]);
 		return idx !== -1 ? idx + 1 : 0;
 	});
-	// themeMenu.newCheckMenu(themeMenu.getMainMenuName(), 'None', options[options.length - 1], () => {
-		// const idx = options.indexOf((file) => {return file === properties.theme[1];});
-		// return (idx !== -1 ? 1 + idx : 0);
-	// });
 	return themeMenu;
 }